requirements.txt
# ─── Core Framework ─────────────────────────────────────────────────────────────
chainlit>=1.1.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.2.0

# ─── LLM & Embeddings ──────────────────────────────────────────────────────────
google-generativeai>=0.5.0
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.3

# ─── Vector Stores ──────────────────────────────────────────────────────────────
chromadb>=0.4.0
faiss-cpu>=1.7.0

# ─── Document Loaders ──────────────────────────────────────────────────────────
pypdf>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0
pandas>=2.0.0
unstructured>=0.10.0
docx2txt>=0.8

# ─── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=23.0.0
tiktoken>=0.5.0
